allow user to choose bacmkapping protocol
